util: Fix parallelization with fail-fast threads
